% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwimp.R
\name{idw}
\alias{idw}
\title{idw function}
\usage{
idw(data, p = 1, k = 1)
}
\arguments{
\item{data}{Data Frame with columns: x,y,z. x: longitudes, y: latitudes, and z precipitation values}

\item{p}{Power of the function, by default is 1}

\item{k}{Number of iterations by default is 1}
}
\value{
Return a Data Frame
}
\description{
Function to fill missing values using idwimp package
}
\examples{

#Example 1:
#precipitation measured in different weather stations
prec = c(259.6716, 262.9709, 288.2743, 307.5199, 328.9999)
#Longitudes of weather stations
lon = c(-77.925, -77.875, -77.825, -77.775, -77.725)
#Latitudes of weather stations
lat = c(2.474999, 2.474999, 2.474999, 2.474999, 2.474999)
#Create gap
precGap = c(259.6716, 262.9709, NA, 307.5199, 328.9999)
df = data.frame(lon,lat,precGap)
filled = idw(df,2,2)
plot(filled$precGap, type="l", col="red")
lines(prec, col="blue")

#Example 2:

data("patiaMissing")
filledGapValues = idw(data = patiaMissing, p= 2, k=2)
#plot(filledGapValues$filledGapValues, type = "l", col = "blue")
#lines(patiaMissing$Prec, col="red")

}
