% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patiaMissing.R
\docType{data}
\name{patiaMissing}
\alias{patiaMissing}
\title{Patia basin precipitation data - 2022-11-05}
\format{
## `patiaMissing`
A data frame with 1036 rows and 3 columns:
\describe{
  \item{x}{Longitudes}
  \item{y}{Latitudes}
  \item{Prec}{Cummulative precipitation}
  ...
}
}
\source{
Acquired from CHIRPS dataset
}
\usage{
patiaMissing
}
\description{
Patia basin is located in southern of Colombia
Report ...
}
\keyword{datasets}
